Добавил визуализацию функций моментов и внес правки | MODEL-1
